refactor app extensions
