+++ app/__init__.py
@@ -4,27 +4,34 @@ import logging
 import mimetypes
 from logging.handlers import RotatingFileHandler
 from flask import Flask, has_request_context, request, session
-from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate
-from flask_socketio import SocketIO
-from flask_assets import Environment, Bundle
-from flask_login import LoginManager
-
 from config import Config
-import os
 
-db = SQLAlchemy(app)
-migrate = Migrate(app, db)
-assets = Environment(app)
-socketio = SocketIO(app,
-                    logging=True,
-                    cors_allowed_origins='*')
-login = LoginManager(app)
 
 def create_app():
     app = Flask(__name__)
     app.config.from_object(Config)
 
+    register_extensions(app)
+    register_blueprints(app)
+
+    return app
+
+
+def register_extensions(app):
+    from app.extensions import db
+    from app.extensions import migrate
+    from app.extensions import assets
+    from app.extensions import socketio
+    from app.extensions import login
+
+    db.init_app(app)
+    migrate.init_app(app, db)
+    assets.init_app(app)
+    socketio.init_app(app)
+    login.init_app(app)
+
+
+def register_blueprints(app):
     from app.api import bp as api_bp
     app.register_blueprint(api_bp, url_prefix='/api')
 
@@ -48,11 +55,14 @@ class RequestFormatter(logging.Formatter):
 
         return super().format(record)
 
+
 formatter = RequestFormatter(
     '[%(asctime)s] User %(user_id)s at %(url)s\n'
     '\t%(message)s'
 )
 
+
+app = create_app()
 
 file_handler = RotatingFileHandler('logs/ringingroom.log','a', 1 * 1024 * 1024, 10)
 file_handler.setFormatter(formatter)
@@ -61,39 +71,15 @@ file_handler.setLevel(logging.ERROR)
 app.logger.addHandler(file_handler)
 app.logger.info('Ringing Room startup')
 
-def log(*args):
-    app.logger.info(' '.join([str(l) for l in args]))
 
-
-# asset bundles
-bundles = {
-
-    'js_landing':   Bundle( 'landing.js',
-                            filters='jsmin',
-                            output='gen/landing.%(version)s.js'),
-
-    'js_rr':        Bundle('ringing_room.js',
-                           'audio.js',
-                           # filters='jsmin',
-                            output='gen/rr.%(version)s.js'),
-    'js_my_towers': Bundle('my_towers.js',
-                           output='gen/my_towers.%(version)s.js'),
-
-    'css_static':   Bundle( 'css/static.css',
-                            output='gen/static.%(version)s.css'),
-
-    'css_rr':   Bundle( 'css/ringing_room.css',
-                         output='gen/rr.%(version)s.css'),
-
-}
 mimetypes.add_type('text/css', '.css')
 mimetypes.add_type('application/javascript', '.js')
 
-assets.register(bundles)
 
-from app.models import *
+from app.models import TowerDict
 
 # make the tower dict
 towers = TowerDict()
+
 
 from app import routes, models, listeners
