+++ app/api/users.py
@@ -1,3 +1,3 @@
-from app.api import bp
+# from app.api import bp
 
 # Routes for the api go here
