+++ app/routes.py
@@ -1,6 +1,7 @@
 from flask import render_template, send_from_directory, abort, flash, redirect, url_for, session, request
 from flask_login import login_user, logout_user, current_user, login_required
-from app import app, towers, log, db
+from app import app, towers
+from app.extensions import db, log
 from app.models import User, UserTowerRelation
 from flask_login import current_user, login_user, logout_user, login_required
 from app.forms import *
@@ -41,7 +42,7 @@ def load_toasts(modal):
 @app.route('/', methods=('GET', 'POST'))
 def index():
     form = LoginForm() if current_user.is_anonymous else None
-    return render_template('landing_page.html', 
+    return render_template('landing_page.html',
                            toasts=load_toasts(modal=False),
                            modals=load_toasts(modal=True),
                            login_form=form)
